requirements.txt
appdirs==1.4.4
astroid==2.9.3
attrs==21.4.0
backports.entry-points-selectable==1.1.1
bandit==1.7.2
black==22.1.0
boto3==1.21.3
botocore==1.24.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coverage==6.3.1
distlib==0.3.4
dnspython==2.2.0
docopt==0.6.2
email-validator==1.1.3
filelock==3.6.0
flake8==4.0.1
Flask==2.0.3
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-WTF==1.0.0
freezegun==1.1.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.7.1
Markdown==3.3.6
MarkupSafe==2.1.0
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pbr==5.8.1
pip-chill==1.0.1
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.7
pytest==7.0.1
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==3.5.0
stripe==2.66.0
toml==0.10.2
tomli==2.0.1
tox==3.24.5
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
watchdog==2.1.6
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
zipp==3.7.0
